fix: accept constrained placeholder trailing-return-type syntax
_has_custom_constrained_auto_param rejected auto f() -> MyConcept
auto; (valid C++20 -- MyConcept auto is a placeholder-type-specifier
usable as a trailing-return-type): the text before the concept name
ends in the "->" from the trailing-return-type arrow, not a
parameter-list boundary or a statement boundary. Reuses the existing
_has_declarator_adjacent_trailing_arrow (already used elsewhere in
this file for the analogous trailing-requires-clause case) to
recognize a genuine declarator-adjacent arrow as an additional valid
position.

Also splits dumper_ast_config.py, which crossed the AI-readiness
2000-line hard cap: the C++20 structural-syntax detection machinery
(concept/requires/consteval/constinit, constrained/abbreviated
parameters, the preprocessor-guard masker) moves to a new sibling
dumper_ast_config_cpp20.py, leaving the AST cache-key and general
C/C++ language-mode detection in dumper_ast_config.py. Pure move, no
behavior change beyond the trailing-return-type fix above.
