+++ abicheck/dumper_ast_config.py
@@ -9,9 +9,7 @@ import hashlib
 import os
 import re
 import shlex
-from dataclasses import dataclass
 from pathlib import Path
-from typing import Literal
 
 from ._compiler_options import has_explicit_std
 from .header_utils import iter_cache_header_files
@@ -131,1742 +129,6 @@ _CPP_ONLY_PATTERNS = [
 # failure hint: here ``extern "C"`` *does* count, because castxml always parses in
 # a C++-ish mode, so an aggregate including an extern "C" header is built as .hpp.
 _CPP_PATTERNS = [_EXTERN_C_PATTERN, *_CPP_ONLY_PATTERNS]
-
-
-# Structural C++20 patterns — concepts and requires-expressions. When any
-# of these appears in a header, castxml must be invoked with a C++20-aware
-# `-std=` flag or it will fail to parse the file. The patterns target the
-# definition site (`concept X = ...`, `requires(...) {`, `template <Foo T>`-
-# style constrained template parameters) rather than uses, so we don't
-# over-trigger. Matching is applied only to *code* text (see
-# ``_find_cpp20_requirements``): preprocessor directive lines, string/char
-# literal contents, and comments are stripped first, so "requires" appearing
-# in a `#error`/`#define` message or a string literal is never mistaken for
-# the C++20 keyword (Codex/false-positive report).
-_CPP20_CONCEPT_PATTERN = re.compile(rb"\bconcept\s+\w+\s*=")  # concept Addable = ...
-# "concept" only became a reserved keyword in C++20 — a pre-C++20 header can
-# legally declare a type literally named "concept" (Codex review, four
-# rounds: a qualified use, an unqualified use, a brace-initialized variable
-# template, and — the general case this pattern exists to catch — a
-# variable template initialized via *any* other expression convertible to
-# that type, e.g. through a converting constructor: ``struct concept {
-# concept(int); }; template<class T> concept C = 1;``). No per-initializer
-# check can ever be complete, so this instead detects the one thing that
-# makes every one of those variants possible: a definition of "concept" as
-# an ordinary type name anywhere in the header. Covers ``union``/``enum``
-# as well as ``struct``/``class`` (Codex review, sixth round) — a
-# ``union concept { ... };`` or ``enum concept { ... };`` shadow is just as
-# legal pre-C++20 as the class-key form and was previously invisible here.
-_CONCEPT_AS_TYPE_NAME_PATTERN = re.compile(
-    rb"\b(?:struct|class|union|enum)\s+concept\b|\busing\s+concept\s*=|\btypedef\b[^;]*\bconcept\s*;"
-)
-# "requires" only became a reserved keyword in C++20 too — a pre-C++20 header
-# can legally declare a type literally named "requires" and reference it in a
-# variable-template declaration (``template<class T> requires value = {};``),
-# whose bare ``requires\s+\w`` shape directly after a template header's
-# closing ``>`` is textually identical to a genuine requires-clause (Codex
-# review). Mirrors ``_CONCEPT_AS_TYPE_NAME_PATTERN``/``concept_type_shadowed``
-# exactly.
-_REQUIRES_AS_TYPE_NAME_PATTERN = re.compile(
-    rb"\b(?:struct|class|union|enum)\s+requires\b|\busing\s+requires\s*=|\btypedef\b[^;]*\brequires\s*;"
-)
-_CPP20_REQUIRES_EXPR_PATTERN = re.compile(
-    rb"\brequires\s*[(\{]"
-)  # requires(T a, T b) { ... }  OR the parameterless requires { ... } form
-# (Codex review: the parameterless form has no parenthesized parameter list —
-# `requires { typename T::value_type; }` — and was previously missed
-# entirely, since the requires-clause pattern below also requires a \w
-# immediately after "requires", which a bare "{" is not.)
-_CPP20_REQUIRES_CLAUSE_PATTERN = re.compile(
-    rb"\brequires\s+\w"
-)  # template<T> requires Foo<T>
-
-# ``consteval``/``constinit`` are new C++20 declaration specifiers, but
-# unlike "concept"/"requires" this is not just a style-vs-risk trade-off:
-# neither was a reserved word before C++20, so a pre-C++20 header can
-# legally use either as an ordinary identifier (``int consteval;``, ``int
-# constinit;`` — declaring a variable with that name). An unconditional
-# bare-keyword match (unlike the deliberately unconditional
-# `_CPP_ONLY_PATTERNS` entries for `constexpr`/`noexcept`/`nullptr`/
-# `override`, which only ever decide "must be C++", not "must be C++20")
-# would force -std=gnu++20 on such a header, where the identifier is no
-# longer usable — actively breaking a header that previously parsed fine
-# (Codex review, second round). Requiring a positive lookahead for
-# whitespace then another identifier-starting character distinguishes a
-# genuine specifier (``consteval int f();``, ``constinit extern int
-# x;`` — always followed by more decl-specifier/declarator content) from
-# the ordinary-identifier shape, where the keyword is the last token of
-# its (simple) declarator, directly followed by ``;``/``,``/``=``/``)``/
-# ``[`` instead.
-#
-# That lookahead alone still can't tell a genuine specifier from a
-# pre-C++20 header that instead declares a *type* literally named
-# "consteval"/"constinit" (``struct consteval {};``) and later
-# references it followed by another decl-specifier or cv-qualifier
-# (``consteval const *p;`` — legal pre-C++20: decl-specifier order is
-# flexible, so this means the same as ``const consteval *p;``) — the
-# textual shape is identical to a genuine ``consteval <type> <name>``
-# declaration (Codex review, third round). Mirrors
-# ``_CONCEPT_AS_TYPE_NAME_PATTERN``/``concept_type_shadowed`` exactly:
-# once a header is confirmed to declare "consteval"/"constinit" as an
-# ordinary type name anywhere, every bare occurrence in that header is
-# ambiguous and treated as non-genuine.
-_CPP20_CONSTEVAL_PATTERN = re.compile(rb"\bconsteval\b(?=\s+[A-Za-z_])")
-_CPP20_CONSTINIT_PATTERN = re.compile(rb"\bconstinit\b(?=\s+[A-Za-z_])")
-_CONSTEVAL_AS_TYPE_NAME_PATTERN = re.compile(
-    rb"\b(?:struct|class|union|enum)\s+consteval\b|\busing\s+consteval\s*=|\btypedef\b[^;]*\bconsteval\s*;"
-)
-_CONSTINIT_AS_TYPE_NAME_PATTERN = re.compile(
-    rb"\b(?:struct|class|union|enum)\s+constinit\b|\busing\s+constinit\s*=|\btypedef\b[^;]*\bconstinit\s*;"
-)
-
-# Constrained template parameters using a *standard-library* concept name in
-# place of ``typename``/``class`` (``template <std::integral T> void f(T);``)
-# — the abbreviated-constraint form the module docstring above already
-# describes but never actually matched (Codex review). Deliberately scoped
-# to the fixed, well-known set of concepts in <concepts>/<iterator>/<ranges>
-# rather than "any bare or qualified identifier in a template parameter
-# list": an arbitrary identifier there is *routinely* a valid pre-C++20
-# non-type template parameter's type (``template<MyEnum E>``,
-# ``template<Traits::value_type V>``), so matching on identifier shape alone
-# would trade this false-negative for a much broader false-positive risk.
-# A `std::`-qualified name from this exact, finite standard list used
-# immediately before a bare template-parameter identifier has no such
-# ambiguity — it is never a plausible NTTP type spelling. <ranges> concepts
-# (Codex review, second round) live under the distinct ``std::ranges::``
-# namespace, matched by a separate pattern below rather than folded into
-# the bare ``std::`` one, since the two prefixes are not interchangeable.
-_CPP20_STD_CONCEPT_NAMES = (
-    rb"same_as|derived_from|convertible_to|common_reference_with|common_with|"
-    rb"integral|signed_integral|unsigned_integral|floating_point|"
-    rb"assignable_from|swappable_with|swappable|destructible|"
-    rb"constructible_from|default_initializable|move_constructible|"
-    rb"copy_constructible|equality_comparable_with|equality_comparable|"
-    rb"totally_ordered_with|totally_ordered|movable|copyable|semiregular|"
-    rb"regular_invocable|regular|invocable|predicate|relation|"
-    rb"strict_weak_order|sortable|mergeable|permutable|indirect_unary_predicate|"
-    rb"indirect_binary_predicate|indirect_equivalence_relation|"
-    rb"indirect_strict_weak_order|indirectly_regular_unary_invocable|"
-    rb"indirectly_unary_invocable|indirectly_readable|indirectly_writable|"
-    rb"indirectly_swappable|indirectly_movable_storable|indirectly_movable|"
-    rb"indirectly_copyable_storable|indirectly_copyable|indirectly_comparable|"
-    rb"weakly_incrementable|incrementable|input_or_output_iterator|"
-    rb"sentinel_for|sized_sentinel_for|input_iterator|output_iterator|"
-    rb"forward_iterator|bidirectional_iterator|random_access_iterator|"
-    rb"contiguous_iterator"
-)
-_CPP20_STD_RANGES_CONCEPT_NAMES = (
-    rb"range|borrowed_range|sized_range|view|input_range|output_range|"
-    rb"forward_range|bidirectional_range|random_access_range|"
-    rb"contiguous_range|common_range|viewable_range|constant_range"
-)
-# Matches just the qualified concept name itself (``std::integral`` /
-# ``std::ranges::range``) — the optional ``<...>`` argument list and what
-# follows it are handled separately by :func:`_has_constrained_param_syntax`,
-# since a naive single-level ``(?:<[^<>]*>)?`` here cannot match a concept
-# argument that itself contains a template-id, e.g.
-# ``std::same_as<std::vector<int>>`` (Codex review, third round).
-_CPP20_CONSTRAINED_PARAM_CONCEPT_PATTERN = re.compile(
-    rb"\bstd::(?:ranges::(?:"
-    + _CPP20_STD_RANGES_CONCEPT_NAMES
-    + rb")|(?:"
-    + _CPP20_STD_CONCEPT_NAMES
-    + rb"))\b"
-)
-# What can follow a constrained template parameter's (optional) concept
-# argument list, before the enclosing template parameter list's next ","
-# or its own closing ">": an optional pack ellipsis, an optional parameter
-# name (anonymous constrained parameters are legal too), and an optional
-# default argument (Codex review — a bare trailing "\w+\s*[,>]" missed
-# both ``template <std::integral T = int>`` and ``template
-# <std::integral... Ts>``). The default's own value is not validated
-# beyond excluding ","/";"/"{"/"}" — a default containing a raw,
-# bracket-free comma (e.g. a function-pointer type) is the one shape this
-# doesn't cover, an acceptable narrow gap given how rare that spelling is
-# in a template-parameter default specifically.
-_CONSTRAINED_PARAM_TAIL_PATTERN = re.compile(
-    rb"(?:\.\.\.)?\s*\w*\s*(?:=\s*[^,;{}]*)?\s*[,>]"
-)
-
-
-def _find_matching_close_angle(text: bytes, open_angle_pos: int) -> int | None:
-    """Return the index of the ``>`` matching the ``<`` at *open_angle_pos*
-    in *text* (tracking nesting), or ``None`` if unbalanced/not found.
-    Mirrors :func:`_find_matching_close_paren` — regex alone cannot match
-    an arbitrarily-nested template-id argument list."""
-    depth = 0
-    for idx in range(open_angle_pos, len(text)):
-        ch = text[idx : idx + 1]
-        if ch == b"<":
-            depth += 1
-        elif ch == b">":
-            depth -= 1
-            if depth == 0:
-                return idx
-    return None
-
-
-def _has_constrained_param_syntax(lookahead: bytes) -> bool:
-    """True if *lookahead* contains a constrained template parameter
-    (``template <std::integral T>``, with the concept argument list — if
-    any — tolerating arbitrary nesting, e.g.
-    ``std::same_as<std::vector<int>>``) or an abbreviated constrained
-    function parameter (``std::integral auto x`` — Codex review: a
-    concept name directly followed by the literal keyword ``auto`` has no
-    other valid pre-C++20 reading, so no disambiguation is needed there).
-    Deliberately scoped to the fixed, well-known set of concepts in
-    <concepts>/<iterator>/<ranges> rather than "any bare or qualified
-    identifier in a template parameter list": an arbitrary identifier
-    there is *routinely* a valid pre-C++20 non-type template parameter's
-    type (``template<MyEnum E>``, ``template<Traits::value_type V>``), so
-    matching on identifier shape alone would trade this false-negative
-    for a much broader false-positive risk. A `std::`-qualified name from
-    this exact, finite standard list has no such ambiguity."""
-    for m in _CPP20_CONSTRAINED_PARAM_CONCEPT_PATTERN.finditer(lookahead):
-        pos = m.end()
-        ws = re.match(rb"\s*", lookahead[pos:])
-        if ws:
-            pos += ws.end()
-        if lookahead[pos : pos + 1] == b"<":
-            close = _find_matching_close_angle(lookahead, pos)
-            if close is None:
-                continue
-            pos = close + 1
-            ws = re.match(rb"\s*", lookahead[pos:])
-            if ws:
-                pos += ws.end()
-        # Whitespace before this point was already consumed above (either
-        # directly after the concept name, or after its <...> argument
-        # list), so what remains needs no further leading \s+ — only the
-        # \b at the concept-name pattern's own end guarantees *some*
-        # non-word separator existed there in the first place.
-        rest = lookahead[pos:]
-        if re.match(rb"auto\b", rest) or _CONSTRAINED_PARAM_TAIL_PATTERN.match(rest):
-            return True
-    return False
-
-
-def _find_enclosing_open_paren(text: bytes, pos: int) -> int | None:
-    """Return the index of the nearest unmatched ``(`` to the left of
-    *pos* in *text*, skipping balanced ``()`` pairs — mirrors
-    :func:`_find_matching_close_paren`'s forward scan, just backward."""
-    depth = 0
-    idx = pos - 1
-    while idx >= 0:
-        ch = text[idx : idx + 1]
-        if ch == b")":
-            depth += 1
-        elif ch == b"(":
-            if depth == 0:
-                return idx
-            depth -= 1
-        idx -= 1
-    return None
-
-
-_TRAILING_ATTRIBUTE_PATTERN = re.compile(rb"\[\[[^\[\]]*\]\]\s*\Z")
-
-
-def _strip_trailing_attributes(prefix: bytes) -> bytes:
-    """Strip zero or more trailing ``[[attr]]``/``[[attr(args)]]``
-    attribute-specifier-seq entries (``[[maybe_unused]]``,
-    ``[[deprecated("msg")]]``, ...) from *prefix* — a standard attribute
-    can precede a parameter's type (``void f([[maybe_unused]] auto x);``),
-    which otherwise leaves the prefix ending in ``]`` instead of the
-    enclosing ``(``/``,`` (Codex review). Narrow gap, accepted: an
-    attribute argument containing a literal ``[``/``]`` (e.g. inside a
-    string) is not unwrapped — vanishingly rare in practice."""
-    prefix = prefix.rstrip()
-    while True:
-        m = _TRAILING_ATTRIBUTE_PATTERN.search(prefix)
-        if not m:
-            return prefix
-        prefix = prefix[: m.start()].rstrip()
-
-
-def _is_lambda_param_list_open_paren(text: bytes, open_paren_pos: int) -> bool:
-    """True if the ``(`` at *open_paren_pos* opens a lambda's parameter
-    list — immediately preceded (skipping whitespace) by the ``]`` that
-    closes a lambda capture list. A generic lambda's ``auto`` parameter
-    (``[](auto x) { ... }``) has been valid since C++14 and must never be
-    mistaken for the C++20-only abbreviated *function* template form."""
-    idx = open_paren_pos - 1
-    while idx >= 0 and text[idx : idx + 1] in b" \t\r\n":
-        idx -= 1
-    return idx >= 0 and text[idx : idx + 1] == b"]"
-
-
-def _is_decltype_open_paren(text: bytes, open_paren_pos: int) -> bool:
-    """True if the ``(`` at *open_paren_pos* is a ``decltype`` specifier's
-    own parentheses — immediately preceded (skipping whitespace) by the
-    keyword ``decltype``. ``decltype(auto)`` (valid since C++14) puts the
-    bare keyword ``auto`` directly inside this ``(``, the identical
-    textual position as a genuine abbreviated parameter's enclosing
-    ``(`` — but it is decltype's own argument, not a parameter list at
-    all (Codex review)."""
-    idx = open_paren_pos - 1
-    while idx >= 0 and text[idx : idx + 1] in b" \t\r\n":
-        idx -= 1
-    end = idx + 1
-    while idx >= 0 and (text[idx : idx + 1].isalnum() or text[idx : idx + 1] == b"_"):
-        idx -= 1
-    return text[idx + 1 : end] == b"decltype"
-
-
-def _has_abbreviated_unconstrained_auto_param(lookahead: bytes) -> bool:
-    """True if *lookahead* contains a bare (unconstrained) ``auto`` used
-    directly as an ordinary function's parameter type (``void f(auto
-    x);``) — the C++20 abbreviated function template form, distinct from
-    the *constrained* form (``std::integral auto x``, handled separately
-    by :func:`_has_constrained_param_syntax`), from a generic lambda's
-    ``auto`` parameter (``[](auto x) { ... }``), which has been valid
-    since C++14 and is excluded via :func:`_is_lambda_param_list_open_paren`
-    (Codex review), and from ``decltype(auto)`` (also valid since C++14,
-    excluded via :func:`_is_decltype_open_paren` — Codex review, second
-    round). Only matches when nothing but an optional cv-qualifier and/or
-    attribute-specifier-seq (``[[maybe_unused]] auto x`` — Codex review,
-    third round) separates ``auto`` from its enclosing ``(``/``,`` — that
-    position is unambiguous: a bare ``auto`` can never be a parameter's
-    default-argument expression or any other operand there, only its
-    type."""
-    for m in re.finditer(rb"\bauto\b", lookahead):
-        prefix = _strip_trailing_declarator_specifiers(lookahead[: m.start()])
-        prefix = _strip_trailing_attributes(prefix)
-        if not prefix:
-            continue
-        last = prefix[-1:]
-        open_pos: int | None
-        if last == b"(":
-            open_pos = len(prefix) - 1
-        elif last == b",":
-            open_pos = _find_enclosing_open_paren(lookahead, len(prefix))
-            if open_pos is None:
-                continue
-        else:
-            continue
-        if not _is_lambda_param_list_open_paren(
-            lookahead, open_pos
-        ) and not _is_decltype_open_paren(lookahead, open_pos):
-            return True
-    return False
-
-
-_QUALIFIED_IDENTIFIER_TAIL_PATTERN = re.compile(rb"(?:[A-Za-z_]\w*::)*[A-Za-z_]\w*\Z")
-
-
-def _has_custom_constrained_auto_param(
-    lookahead: bytes, prev_nonblank_code: bytes = b""
-) -> bool:
-    """True if *lookahead* contains an abbreviated function parameter
-    constrained by a project-defined (non-``std::``) concept
-    (``void f(MyConcept auto x);``) — Codex review. Unlike a bare
-    identifier directly inside a ``template<...>`` parameter list
-    (routinely a valid pre-C++20 non-type template parameter's type, e.g.
-    ``template<MyEnum E>`` — see :func:`_has_constrained_param_syntax`'s
-    docstring for why that shape stays scoped to the finite ``std::``
-    list), an identifier immediately followed by the keyword ``auto`` in a
-    parameter's decl-specifier position has no valid pre-C++20 reading at
-    all: a decl-specifier-seq can carry at most one type-determining
-    specifier, and ``auto`` — always a keyword, never usable as an
-    ordinary identifier the way ``concept``/``requires``/``consteval``/
-    ``constinit`` are pre-C++20 — is itself one, so juxtaposing it with an
-    unrelated type name is exclusively the C++20 constrained-placeholder
-    syntax regardless of which concept is named. Reuses the same
-    unambiguous-boundary check as :func:`_has_abbreviated_unconstrained_auto_param`
-    (nothing but the identifier separates ``auto`` from its enclosing
-    ``(``/``,``); a leading cv-qualifier before the identifier
-    (``const MyConcept auto& x``, Codex review) is stripped via
-    :func:`_strip_trailing_declarator_specifiers` first — reused here for
-    its trailing-word-popping loop, even though this call site's prefix
-    isn't itself a trailing-specifier position — so the boundary check
-    that follows (``(``/``,``, or a genuine statement boundary for the
-    return-type shape) still sees the real enclosing punctuation once
-    ``const``/``volatile`` is out of the way.
-
-    Unlike :func:`_has_abbreviated_unconstrained_auto_param`, a lambda's
-    parameter list is deliberately *not* excluded here (Codex review):
-    a *bare* ``auto`` lambda parameter (``[](auto x) {}``) has been valid
-    since C++14, but a lambda parameter *constrained* by a concept
-    (``[](MyConcept auto x) {}``) is exactly as C++20-only as the
-    equivalent ordinary-function form — lambdas' call operators gained
-    the same abbreviated-function-template treatment in C++20. Excluding
-    lambda parameter lists here would silently miss the only C++20 signal
-    in a header whose lone use of the feature is a constrained generic
-    lambda.
-
-    A constrained placeholder *return type* (``MyConcept auto f();``,
-    Codex review) gets the identical unambiguous decl-specifier-seq
-    reasoning — it counts whenever nothing but whitespace separates the
-    identifier from a genuine statement boundary: either earlier content
-    on the same logical line ending in ``;``/``{``/``}``, or (when the
-    identifier opens the line entirely) *prev_nonblank_code*'s last
-    non-blank code ending the same way, or no preceding code at all (the
-    very first declaration in the scanned content). *prev_nonblank_code*
-    defaults to empty for callers that don't track it (matching this
-    module's other lookahead-only helpers) — the return-type check simply
-    never fires for a caller that doesn't pass it, the same conservative
-    default as any other unattempted shape here."""
-    for m in re.finditer(rb"\bauto\b", lookahead):
-        prefix = _strip_trailing_declarator_specifiers(lookahead[: m.start()])
-        prefix = _strip_trailing_attributes(prefix)
-        ident_match = _QUALIFIED_IDENTIFIER_TAIL_PATTERN.search(prefix)
-        if not ident_match:
-            continue
-        before_ident = prefix[: ident_match.start()].rstrip()
-        before_ident = _strip_trailing_declarator_specifiers(before_ident)
-        last = before_ident[-1:]
-        open_pos: int | None
-        if last == b"(":
-            open_pos = len(before_ident) - 1
-        elif last == b",":
-            open_pos = _find_enclosing_open_paren(lookahead, len(before_ident))
-            if open_pos is None:
-                continue
-        else:
-            open_pos = None
-        if open_pos is not None:
-            if not _is_decltype_open_paren(lookahead, open_pos):
-                return True
-            continue
-        # Not inside a parameter list at all — only a genuine statement
-        # boundary right before the identifier (on this line, or on the
-        # last preceding non-blank code when the identifier opens the
-        # line) makes this a declaration-start return-type position
-        # rather than some other expression context.
-        if before_ident:
-            if last not in _REQUIRES_STATEMENT_BOUNDARY_CHARS:
-                continue
-        else:
-            prev = _strip_trailing_declarator_specifiers(prev_nonblank_code.rstrip())
-            if prev and prev[-1:] not in _REQUIRES_STATEMENT_BOUNDARY_CHARS:
-                continue
-        return True
-    return False
-
-
-# "requires" only became a reserved keyword in C++20 — any earlier standard
-# allows it as an ordinary identifier, e.g. ``bool requires(int x) { ... }``
-# (a declaration) or ``requires(1);`` (a call), both real uses of a
-# function literally named "requires". Forcing -std=gnu++20 on such a
-# header would break it, since the identifier is no longer usable there
-# (Codex review, two rounds: the declaration case, then the call/
-# expression-statement case). The only way "requires(" is preceded by a
-# bare word (just whitespace, no operator) in *genuine* C++20 usage is a
-# handful of expression-introducing keywords (return/throw/co_return);
-# every other preceding identifier can only be a declaration/call using
-# "requires" as a plain pre-C++20 name — juxtaposing two bare identifiers
-# with nothing but whitespace between them is not valid C++ in any other
-# production. Likewise, "requires(" at the very start of a statement (
-# preceded by nothing, or by a statement-boundary "{"/"}"/";") can only be
-# a call-as-statement using the plain pre-C++20 name — a genuine
-# requires-expression or requires-clause is always itself a sub-expression
-# (an operand), never a bare statement by construction of this detector's
-# own trigger (there is no standalone top-level "requires ...;" construct
-# in C++20 outside of being part of a larger declaration/expression).
-_REQUIRES_EXPR_SAFE_PRECEDING_WORDS = frozenset({b"return", b"throw", b"co_return"})
-_REQUIRES_STATEMENT_BOUNDARY_CHARS = frozenset({b"{", b"}", b";"})
-_TRAILING_IDENTIFIER_PATTERN = re.compile(rb"([A-Za-z_]\w*)\Z")
-
-
-def _find_matching_close_paren(text: bytes, open_paren_pos: int) -> int | None:
-    """Return the index of the ``)`` matching the ``(`` at *open_paren_pos*
-    in *text* (tracking nesting), or ``None`` if unbalanced/not found."""
-    depth = 0
-    for idx in range(open_paren_pos, len(text)):
-        ch = text[idx : idx + 1]
-        if ch == b"(":
-            depth += 1
-        elif ch == b")":
-            depth -= 1
-            if depth == 0:
-                return idx
-    return None
-
-
-def _requires_match_has_body(lookahead: bytes, match: re.Match[bytes]) -> bool:
-    """True if the requires(...)/requires{ *match* is confirmed to carry a
-    requirements body: the parameterless form always does (the matched
-    ``{`` **is** the body); the parenthesized form only does when its
-    matching ``)`` is immediately followed by ``{``. A plain call to a
-    pre-C++20 "requires" function has no such body — just ``;`` or another
-    token after the closing paren."""
-    matched_char = lookahead[match.end() - 1 : match.end()]
-    if matched_char == b"{":
-        return True
-    close = _find_matching_close_paren(lookahead, match.end() - 1)
-    if close is None:
-        return False
-    return lookahead[close + 1 :].lstrip().startswith(b"{")
-
-
-_TRAILING_DECLARATOR_SPECIFIER_WORDS = frozenset(
-    {b"const", b"volatile", b"noexcept", b"override", b"final"}
-)
-
-
-def _strip_trailing_declarator_specifiers(prefix: bytes) -> bytes:
-    """Strip zero or more trailing cv/ref-qualifiers and specifiers
-    (``const``, ``volatile``, ``noexcept``, ``override``, ``final``,
-    ``&``, ``&&``) from *prefix* (Codex review). A trailing requires-clause
-    can follow any number of these after a function's declarator (
-    ``void f(T) const noexcept requires C<T>;``) — stripping them lets the
-    caller still trace the prefix back to the declarator's own closing
-    ``)`` regardless of how many specifiers sit in between. Safe to strip
-    unconditionally: none of these words/operators can appear as a bare
-    trailing token in *any* other C++ construct without being part of a
-    declarator's trailing specifier sequence."""
-    prefix = prefix.rstrip()
-    changed = True
-    while changed:
-        changed = False
-        if prefix.endswith(b"&&"):
-            prefix = prefix[:-2].rstrip()
-            changed = True
-        elif prefix.endswith(b"&"):
-            prefix = prefix[:-1].rstrip()
-            changed = True
-        else:
-            m = _TRAILING_IDENTIFIER_PATTERN.search(prefix)
-            if m is not None and m.group(1) in _TRAILING_DECLARATOR_SPECIFIER_WORDS:
-                prefix = prefix[: m.start()].rstrip()
-                changed = True
-    return prefix
-
-
-def _has_declarator_adjacent_trailing_arrow(prefix: bytes) -> bool:
-    """True if *prefix* contains a trailing-return-type arrow (``->``)
-    that itself directly follows (after stripping cv/ref/noexcept
-    specifiers) a function declarator's closing ``)`` — the ``->
-    ReturnType`` shape of ``auto f(T) -> ReturnType``.
-
-    A bare substring search for ``->`` anywhere in *prefix* also matches
-    an unrelated member-access expression earlier in the *same*
-    statement/expression, not just a different one (the statement-
-    boundary check catches that case, but not this one): ``int
-    requires(int); return p->m + requires(1);`` — a plain pre-C++20 call
-    to a function named "requires", added to a member-access result —
-    was wrongly classified genuine because of the ``->`` in ``p->m``, with
-    no statement boundary between it and "requires" (Codex review, fifth
-    round). Walking every ``->`` occurrence right-to-left (rather than
-    just checking substring membership) also correctly finds a nested
-    arrow inside the return type itself (a rare ``decltype(a->b)`` return
-    type) even when it is not the rightmost occurrence.
-
-    Residual, accepted ambiguity: a function *call* immediately followed
-    by member access (``getObj()->m``) has the identical ``...)  ->...``
-    shape as a genuine declarator, and this check cannot tell them apart
-    without real parsing — the same "impossible to bound generically"
-    trade-off already accepted for the return-type expression itself."""
-    for m in re.finditer(re.escape(b"->"), prefix):
-        before = _strip_trailing_declarator_specifiers(prefix[: m.start()])
-        if before.endswith(b")"):
-            return True
-    return False
-
-
-def _looks_like_requires_declarator(
-    lookahead: bytes, match: re.Match[bytes], prev_nonblank_code: bytes
-) -> bool:
-    """True if the requires-expression candidate at *match* in *lookahead*
-    looks like an ordinary pre-C++20 use of "requires" as a plain
-    identifier — either immediately preceded (skipping only whitespace) by
-    a bare identifier that isn't one of the few keywords that can
-    legitimately introduce a requires-expression as an operand (the
-    declaration/call-with-preceding-name case), preceded by nothing but a
-    statement boundary (the bare call-as-statement case), or preceded by
-    ``.``/``->``/``::`` (a member/qualified-name access — "requires" the
-    C++20 keyword is never looked up that way) — rather than the C++20
-    keyword.
-
-    When nothing at all precedes the candidate on its own logical line, a
-    genuine parenthesized requires-clause continuing a ``template<...>``
-    header from the *previous* line (``template<class T>\\nrequires
-    (sizeof(T) > 4)\\nvoid f(T);``) looks identical to a bare call-as-
-    statement at this point, so that case falls back to *prev_nonblank_code*
-    the same way :func:`_looks_like_genuine_concept` does — but only after
-    confirming *prev_nonblank_code* is not itself a *different*, unrelated
-    statement (a leading statement-boundary character in it means the
-    genuine-continuation shape above cannot apply at all): otherwise a
-    stray ``->``/``)``/``>`` anywhere earlier in that previous logical
-    line — which can hold more than one statement — was picked up by the
-    same unscoped-substring bug as :func:`_looks_like_genuine_requires_clause`
-    (Codex review, fourth round).
-
-    A safe preceding word (return/throw/co_return) is necessary but not
-    sufficient: ``return requires(1);`` — a plain call to a pre-C++20
-    "requires" function — is just as syntactically valid there as a real
-    ``return requires(T t) { t.foo(); };``. Only the latter carries a
-    requirements body, so that case additionally confirms one before
-    accepting (Codex review). The same ambiguity applies whenever
-    "requires" is preceded by an operator/punctuation rather than a bare
-    identifier at all — ``if (requires(1)) ...``, ``!requires(1)``,
-    ``x = requires(1);`` — a plain call is just as valid there as a genuine
-    requires-expression used as an operand, so that case also falls back to
-    the body check rather than assuming genuine (Codex review, second
-    round).
-
-    A requires-*clause* with a parenthesized constraint directly
-    continuing its own ``template<...>`` header on the *same* line
-    (``template<class T> requires (sizeof(T) > 4) void f(T);``) has no
-    trailing ``{`` body — a clause is not an expression — so without this
-    check it would be misjudged as a plain pre-C++20 call by the body-check
-    fallback below, the same way a bare trailing ``requires`` on its own
-    line already falls back to *prev_nonblank_code* for this. Checked only
-    *after* the member-access exclusion, since ``->`` itself ends in ``>``
-    and must not be mistaken for a template header's closing angle bracket
-    (Codex review, third round; regression caught locally before commit).
-    Mirrors :func:`_looks_like_genuine_concept`'s identical same-line
-    check.
-
-    A parenthesized requires-clause can equally *trail* a function's
-    declarator (``void f(T) requires (sizeof(T) > 4);``, or after cv/ref/
-    ``noexcept`` specifiers — ``void f(T) const noexcept requires
-    (sizeof(T) > 4);``, or a trailing return type — ``auto f(T) -> int
-    requires (sizeof(T) > 4);``), which the body-check fallback below
-    cannot recognize on its own — a clause has no body to confirm (Codex
-    review, two rounds). :func:`_strip_trailing_declarator_specifiers`
-    traces the prefix back through cv/ref/``noexcept`` specifiers to the
-    parameter list's own closing ``)``; a ``->`` anywhere in what remains
-    signals a trailing return type instead — the same unambiguous
-    positional signals already used in
-    :func:`_looks_like_genuine_requires_clause`."""
-    prefix = lookahead[: match.start()].rstrip()
-    if not prefix:
-        prev = _strip_trailing_declarator_specifiers(prev_nonblank_code.rstrip())
-        if prev[-1:] in _REQUIRES_STATEMENT_BOUNDARY_CHARS:
-            return True
-        return not (
-            prev.endswith(b">")
-            or prev.endswith(b")")
-            or _has_declarator_adjacent_trailing_arrow(prev)
-        )
-    if prefix[-1:] in _REQUIRES_STATEMENT_BOUNDARY_CHARS:
-        return True
-    if prefix.endswith(b".") or prefix.endswith(b"->") or prefix.endswith(b"::"):
-        return True
-    stripped = _strip_trailing_declarator_specifiers(prefix)
-    if (
-        stripped.endswith(b">")
-        or stripped.endswith(b")")
-        or _has_declarator_adjacent_trailing_arrow(stripped)
-    ):
-        return False
-    m = _TRAILING_IDENTIFIER_PATTERN.search(prefix)
-    if m is not None and m.group(1) not in _REQUIRES_EXPR_SAFE_PRECEDING_WORDS:
-        return True
-    return not _requires_match_has_body(lookahead, match)
-
-
-def _looks_like_genuine_concept(
-    lookahead: bytes,
-    match: re.Match[bytes],
-    prev_nonblank_code: bytes,
-    concept_type_shadowed: bool,
-) -> bool:
-    """True only if the concept-declaration candidate is actually preceded
-    by a ``template<...>`` header's closing ``>`` — either earlier on the
-    same (possibly lookahead-joined) line, or as the last thing on the
-    previous non-blank code line when "concept" itself starts this one. A
-    concept-name is always declared bare, directly after its own
-    ``template<...>`` header, so requiring this positive signal (rather
-    than merely excluding a ``::`` prefix) is what actually distinguishes a
-    genuine declaration from "concept" being used as an ordinary pre-C++20
-    identifier anywhere else in a statement (Codex review: excluding only
-    ``::`` still missed a plain, unqualified pre-C++20 use like
-    ``static concept C = {};``).
-
-    Even with a preceding template header, "concept" only became a
-    reserved keyword in C++20 — a pre-C++20 header can legally declare a
-    type literally named "concept" and use it in an ordinary *variable
-    template* (``template<class T> concept C = {};``, valid since
-    C++14), which has the identical textual shape as a genuine concept
-    definition (Codex review, several rounds). No per-initializer-shape
-    check can be complete — the variable template's initializer can be
-    *any* expression convertible to the shadowing type, not just a
-    brace-init-list (e.g. ``struct concept { concept(int); }; ...
-    concept C = 1;`` via a converting constructor) — so *whenever this
-    header defines "concept" as a real type anywhere*
-    (``concept_type_shadowed``), every bare ``concept NAME = ...`` match
-    in it is ambiguous and rejected outright, regardless of what follows
-    "="."""
-    if concept_type_shadowed:
-        return False
-    same_line_prefix = lookahead[: match.start()].rstrip()
-    if same_line_prefix.endswith(b">"):
-        return True
-    if not same_line_prefix:
-        return prev_nonblank_code.rstrip().endswith(b">")
-    return False
-
-
-def _looks_like_genuine_requires_clause(
-    lookahead: bytes,
-    match_start: int,
-    prev_nonblank_code: bytes,
-    requires_type_shadowed: bool,
-) -> bool:
-    """True only if the requires-*clause* candidate (the bare, non-
-    parenthesized ``requires Foo<T>`` form matched by
-    ``_CPP20_REQUIRES_CLAUSE_PATTERN``) is actually preceded by a
-    ``template<...>`` header's closing ``>`` — mirrors
-    :func:`_looks_like_genuine_concept` exactly, for the same reason: a
-    plain pre-C++20 declaration using "requires" as an ordinary type/
-    variable name (``struct requires {}; requires value;`` — declaring a
-    variable of type "requires") has the identical bare
-    ``requires\\s+\\w`` shape as a genuine clause, and was previously
-    accepted unconditionally by this branch with no declarator check at
-    all (Codex review). Unlike the parenthesized/brace-delimited
-    requires-expression form (:func:`_looks_like_requires_declarator`), a
-    clause has no body to confirm, so the *only* positive signal
-    available is the preceding template header — exactly the same
-    positive-signal-required design already used for ``concept``.
-
-    That template-header check alone still can't tell a genuine clause
-    from a pre-C++20 header that declares a *type* literally named
-    "requires" and uses it as a variable template's type
-    (``template<class T> requires value = {};`` — the template header's
-    closing ``>`` directly precedes "requires" either way) — Codex
-    review, sixth round. Once *requires_type_shadowed* confirms "requires"
-    names a real type anywhere in the header, every candidate here is
-    ambiguous and treated as non-genuine, mirroring
-    ``concept_type_shadowed``.
-
-    A *trailing* requires-clause following a function's declarator (
-    ``template<class T> void f(T) requires std::integral<T>;``) is
-    equally genuine, signaled by the prefix ending in the parameter
-    list's closing ``)`` instead — Codex review. This is unambiguous:
-    nothing but a trailing specifier (cv/ref-qualifier, ``noexcept``, a
-    requires-clause, ...) can follow a function declarator's ``)``
-    before the terminating ``;``/``{`` in *any* C++ grammar, pre-C++20
-    included — there is no production for a second, unrelated statement
-    beginning right there with no separator. Any number of such
-    specifiers (``void f(T) const noexcept requires C<T>;``) can sit
-    between the ``)`` and the clause — Codex review, second round —
-    traced back via :func:`_strip_trailing_declarator_specifiers`.
-
-    A trailing return type (``auto f(T) -> int requires C<T>;``) is the
-    same shape once more removed: whatever sits between ``->`` and
-    ``requires`` is the return-type expression, itself impossible to
-    bound generically, but its mere presence is enough — a bare
-    ``requires IDENTIFIER`` directly following *any* token with no
-    separator is, by the same invariant as the ``)``/``>`` cases, only
-    ever valid pre-C++20 as a two-identifier ``Type Name;`` declaration,
-    and that shape requires "requires" to be preceded by *nothing but*
-    the type name — never by a ``->`` (which only ever introduces a
-    trailing return type or a member access, neither of which can
-    itself be the "type name" half of such a declaration) — Codex
-    review, third round.
-
-    The ``->`` check must stay scoped to the *current* statement: a bare
-    substring search across the whole same-line prefix picks up an
-    unrelated earlier statement's ``->`` too (``auto x = p->m; requires
-    value;`` — an ordinary pre-C++20 declaration of ``value`` with type
-    "requires" — was wrongly classified genuine by an arrow belonging to
-    the *previous* statement, forcing ``-std=gnu++20`` and rejecting the
-    otherwise-valid header). A statement boundary (``;``/``{``/``}``)
-    directly preceding "requires" can never be a genuine clause's
-    continuation, so it is excluded first, mirroring
-    :func:`_looks_like_requires_declarator`'s identical check — Codex
-    review, fourth round. That statement-boundary check alone still
-    leaves the *same*-statement case open — an unrelated ``->`` earlier
-    in the same expression (``return p->m + requires(1);``, no statement
-    boundary between them) — so the arrow check itself now requires the
-    ``->`` to be declarator-adjacent via
-    :func:`_has_declarator_adjacent_trailing_arrow` rather than a bare
-    substring search — Codex review, fifth round."""
-    if requires_type_shadowed:
-        return False
-    same_line_prefix = _strip_trailing_declarator_specifiers(
-        lookahead[:match_start].rstrip()
-    )
-    if same_line_prefix[-1:] in _REQUIRES_STATEMENT_BOUNDARY_CHARS:
-        return False
-    if (
-        same_line_prefix.endswith(b">")
-        or same_line_prefix.endswith(b")")
-        or _has_declarator_adjacent_trailing_arrow(same_line_prefix)
-    ):
-        return True
-    if not same_line_prefix:
-        prev = _strip_trailing_declarator_specifiers(prev_nonblank_code.rstrip())
-        if prev[-1:] in _REQUIRES_STATEMENT_BOUNDARY_CHARS:
-            return False
-        return (
-            prev.endswith(b">")
-            or prev.endswith(b")")
-            or _has_declarator_adjacent_trailing_arrow(prev)
-        )
-    return False
-
-
-_STRING_LITERAL_PATTERN = re.compile(rb'"(?:\\.|[^"\\\n])*"')
-_CHAR_LITERAL_PATTERN = re.compile(rb"'(?:\\.|[^'\\\n])*'")
-# C++11 raw string literal: [prefix]R"delim(...)delim" — the standard
-# permits an optional encoding prefix (u8, u, U, L) directly before the R,
-# e.g. u8R"(...)"; without it, "\bR" never matches after "u8"/"u"/"U"/"L"
-# since both characters are \w (no boundary between them), leaving a
-# prefixed raw string completely unstripped (Codex review). The delimiter
-# (d-char-sequence) grammar permits any basic-source character except
-# whitespace, parentheses, and backslash — not just identifier characters
-# (a delimiter like "tag-" is valid and was missed by an earlier,
-# identifier-only version of this pattern (Codex review) — matching the
-# exclusion directly is simpler and more complete than enumerating every
-# permitted punctuation character. Not handled by _STRING_LITERAL_PATTERN
-# (only ordinary "..." literals) or by the plain-comment stripper, so its
-# body was otherwise scanned as ordinary code: text that merely *looks*
-# like a requires-expression/concept inside a raw string would force
-# -std=gnu++20 unnecessarily — worse once a multi-line construct can span
-# into a raw string's later lines too. DOTALL so the (non-greedy) body can
-# span newlines.
-_RAW_STRING_LITERAL_PATTERN = re.compile(
-    rb'\b(?:u8|u|U|L)?R"([^\s()\\]{0,16})\((?:.*?)\)\1"', re.DOTALL
-)
-
-
-def _strip_raw_strings(content: bytes) -> bytes:
-    """Blank C++11 raw string literals entirely (delimiter and body alike),
-    preserving embedded newlines so line numbers reported for code after a
-    multi-line raw string stay accurate (mirrors the block-comment
-    stripper's newline-preserving approach)."""
-    return _RAW_STRING_LITERAL_PATTERN.sub(
-        lambda m: b"\n" * m.group(0).count(b"\n"), content
-    )
-
-
-def _strip_literals(line: bytes) -> bytes:
-    """Blank out string/char literal contents.
-
-    Prevents a keyword that only appears *inside* a string (e.g. an error
-    message like ``"Foo requires Base"``) from being mistaken for C++
-    structural syntax.
-    """
-    line = _STRING_LITERAL_PATTERN.sub(b'""', line)
-    line = _CHAR_LITERAL_PATTERN.sub(b"''", line)
-    return line
-
-
-# Newline-tolerant variants of the two patterns above, for use ONLY on a
-# chunk that has already been through _iter_logical_lines: that step splices
-# away a backslash-newline continuation, embedding a literal "\n" exactly
-# where the continuation was — so an ordinary string literal like
-# ``"requires \`` + newline + ``{ ... }"`` (a real, if archaic, C/C++
-# feature) arrives with the keyword and brace on either side of an embedded
-# newline the plain patterns above deliberately refuse to cross (bounding an
-# unterminated-literal mismatch to one line is the whole point there). A
-# single already-joined logical line has no such risk — any embedded
-# newline in it is a genuine continuation, not a boundary into unrelated
-# code — so it is safe to let ``.`` span it here (Codex review: this is what
-# let requires/concept text trapped inside a continued string literal reach
-# the structural pattern match).
-_JOINED_STRING_LITERAL_PATTERN = re.compile(rb'"(?:\\.|[^"\\])*"', re.DOTALL)
-_JOINED_CHAR_LITERAL_PATTERN = re.compile(rb"'(?:\\.|[^'\\])*'", re.DOTALL)
-
-
-def _strip_literals_joined(line: bytes) -> bytes:
-    """Like :func:`_strip_literals`, but tolerant of an embedded newline —
-    use only on output from :func:`_iter_logical_lines`."""
-    line = _JOINED_STRING_LITERAL_PATTERN.sub(b'""', line)
-    line = _JOINED_CHAR_LITERAL_PATTERN.sub(b"''", line)
-    return line
-
-
-def _strip_literals_crossing_continuations(content: bytes) -> bytes:
-    """Like :func:`_strip_literals`, but — unlike it — a literal that spans a
-    backslash-newline continuation is still fully blanked, not left behind
-    because the plain patterns refuse to cross the embedded newline.
-
-    Safe to call directly on whole-file *content* (unlike
-    :func:`_strip_literals_joined`, which additionally tolerates crossing an
-    *unrelated* later line and so is only safe on a single already-joined
-    logical line): ``\\.`` under ``re.DOTALL`` already consumes a
-    continuation's ``\\<newline>`` pair as one escaped character, so the
-    match still ends at the literal's real closing quote rather than
-    wandering into unrelated later lines. Each replacement preserves the
-    literal's embedded newline count (mirrors :func:`_strip_raw_strings`) so
-    line numbers reported for code that follows a continued literal stay
-    accurate — the plain ``_strip_literals_joined`` replacement (a bare
-    ``""``/``''``) would otherwise silently swallow those newlines (Codex
-    review: a shadow-name scan run before comment-stripping needs a
-    continuation-spanning literal fully blanked, or a fake type name like
-    ``struct concept {};`` trapped inside one leaks through and wrongly
-    shadows a genuine C++20 declaration elsewhere in the header).
-    """
-    content = _JOINED_STRING_LITERAL_PATTERN.sub(
-        lambda m: b'""' + b"\n" * m.group(0).count(b"\n"), content
-    )
-    content = _JOINED_CHAR_LITERAL_PATTERN.sub(
-        lambda m: b"''" + b"\n" * m.group(0).count(b"\n"), content
-    )
-    return content
-
-
-def _iter_logical_lines(content: bytes) -> list[tuple[int, bytes]]:
-    """Split *content* into ``(1-based start line, logical line)`` pairs.
-
-    Backslash-newline continuations are joined into a single logical line so
-    a ``#define``/``#error`` directive spanning multiple physical lines is
-    classified as one directive rather than leaking its continuation lines
-    into ordinary code scanning.
-    """
-    physical = content.split(b"\n")
-    logical: list[tuple[int, bytes]] = []
-    start_no = 1
-    buf: list[bytes] = []
-    for i, raw in enumerate(physical, start=1):
-        line = raw.rstrip(b"\r")
-        if not buf:
-            start_no = i
-        if line.endswith(b"\\"):
-            buf.append(line[:-1])
-            continue
-        buf.append(line)
-        logical.append((start_no, b"\n".join(buf)))
-        buf = []
-    if buf:
-        logical.append((start_no, b"\n".join(buf)))
-    return logical
-
-
-def _is_preprocessor_directive(line: bytes) -> bool:
-    return re.match(rb"^\s*#", line) is not None
-
-
-_PP_IF_OPEN_PATTERN = re.compile(rb"^[ \t]*#[ \t]*(?:if|ifdef|ifndef)\b")
-# A parenthesized literal (``#if (0)``, ``#elif (1)``) is the same
-# permanently-false/-true guard as the bare spelling — valid C/C++
-# preprocessor syntax, just with redundant grouping parens around the
-# whole condition (Codex review, ninth round). Both parenthesized and bare
-# forms are accepted; a stray unbalanced paren is invalid C anyway, so
-# matching only the balanced pair here is sufficient.
-_PP_LITERAL_ZERO = rb"(?:\([ \t]*(?:0+|false)[ \t]*\)|(?:0+|false))"
-_PP_LITERAL_TRUE = rb"(?:\([ \t]*(?:1|true)[ \t]*\)|(?:1|true))"
-_PP_IF_ZERO_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*if[ \t]+" + _PP_LITERAL_ZERO + rb"[ \t]*$"
-)
-# ``#if 1``/``#if true`` (Codex review, tenth round): an unconditionally
-# *true* opening arm was never recognized as starting a trackable chain at
-# all (only ``#if 0``/feature-guard openings were), so a dead ``#else``/
-# ``#elif`` sibling of a genuinely-true ``#if 1`` fell through to plain
-# pass-through scanning right along with the live arm -- a real C++20
-# construct sitting only in that dead sibling would then force
-# ``-std=gnu++20`` onto a header whose *active* code never needed it, and
-# could break that active code outright if it uses a soon-to-be-reserved
-# word (``consteval``, ``concept``, ...) as an ordinary identifier. Track
-# it like ``#elif 1``/``#elif true``: the opening arm itself stays
-# unmasked (it's live), but the chain is immediately ``settled`` so every
-# later sibling arm is masked as unreachable.
-_PP_IF_TRUE_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*if[ \t]+" + _PP_LITERAL_TRUE + rb"[ \t]*$"
-)
-_PP_ELSE_PATTERN = re.compile(rb"^[ \t]*#[ \t]*else\b")
-_PP_ELIF_PATTERN = re.compile(rb"^[ \t]*#[ \t]*elif\b")
-_PP_ELIF_ZERO_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*elif[ \t]+" + _PP_LITERAL_ZERO + rb"[ \t]*$"
-)
-_PP_ELIF_TRUE_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*elif[ \t]+" + _PP_LITERAL_TRUE + rb"[ \t]*$"
-)
-_PP_ENDIF_PATTERN = re.compile(rb"^[ \t]*#[ \t]*endif\b")
-# A guard on __cplusplus (or a standard feature-test macro, __cpp_*) is
-# self-consistent under *any* -std= this heuristic ends up choosing — the
-# guard's own condition is what makes its content reachable or not, driven
-# by that same choice, so the content behind it must never be allowed to
-# drive the choice itself (Codex review). Masking it exactly like #if 0
-# breaks that circularity: whichever std gets picked, the guard evaluates
-# consistently and the header parses either way.
-_PP_IF_CPLUSPLUS_GUARD_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*if[ \t]+.*\b(?:__cplusplus|__cpp_\w+)\b"
-)
-_PP_ELIF_CPLUSPLUS_GUARD_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*elif[ \t]+.*\b(?:__cplusplus|__cpp_\w+)\b"
-)
-# ``#if __cplusplus < N`` (or ``<=``) is the same "not yet at this
-# dialect" polarity as ``#ifndef __cpp_x`` (Codex review, fifteenth
-# round), not the ``__cplusplus >= N`` polarity the general guard pattern
-# above assumes: the general pattern masks the *guarded* arm and trusts
-# the ``#else`` uniformly, which is only correct when the guarded arm is
-# the feature-*present* content (``>=``/``>``). For a less-than
-# comparison the guarded arm is the pre-C++20-safe fallback and the
-# ``#else`` is the circular, feature-present content instead — reusing
-# the general pattern's treatment here masked the safe pre-C++20 arm and
-# trusted the circular ``#else``, so a header written "if older than
-# C++20, do X; else do Y" forced ``-std=gnu++20`` purely because Y was
-# there, which could then break an unrelated, genuinely pre-C++20 use of
-# the same word as an ordinary identifier in the masked-away X arm.
-#
-# Restricted to the exact C++20-boundary literals (Codex review,
-# seventeenth round): the inverted polarity only holds when the
-# threshold *is* C++20 (``202002L``/``201703L``, the two ISO values that
-# both mean "not yet C++20" depending on which operator's used). A
-# less-than check against an *older* standard's threshold
-# (``__cplusplus < 201103L``, "before C++11") is not about the C++20
-# decision at all — there the guarded arm is the ancient-dialect
-# fallback that's realistically never reached (castxml/clang always
-# default to well past C++11), and the ``#else`` is the one that's
-# practically always active, so the *general* (mask-guarded/trust-else)
-# treatment is already correct for it. Blanket-inverting every
-# less-than comparison regardless of threshold masked that always-active
-# ``#else`` too, hiding a genuine, unconditional C++20 construct there.
-# Deliberately narrow otherwise (matching this file's
-# incremental-per-reported-case scope, not a full expression parser):
-# reversed operand order (``202002L > __cplusplus``), a combined
-# condition (``__cplusplus < 202002L && defined(FOO)``), and other
-# comparison operators (``==``, ``!=``) are not attempted here and keep
-# falling through to the general pattern above, the same conservative
-# default an entirely unrecognized condition gets.
-_PP_CPLUSPLUS_LESS_THAN_CPP20_TAIL = (
-    rb"__cplusplus[ \t]*(?:<[ \t]*202002L?|<=[ \t]*201703L?)"
-)
-_PP_IF_CPLUSPLUS_LESS_THAN_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*if[ \t]+" + _PP_CPLUSPLUS_LESS_THAN_CPP20_TAIL + rb"[ \t]*$"
-)
-# Same fix, ``#elif`` spelling (Codex review, sixteenth round): a
-# less-than fallback can just as well be written as a later arm in a
-# chain that starts with a disabled/unrelated arm
-# (``#if 0 ... #elif __cplusplus < N ... #else ... #endif``). Treated
-# exactly like ``#elif 1``/``#elif true`` for masking purposes: it's the
-# trustworthy arm once reached, so it settles the chain the same way —
-# masked if an earlier arm already settled it, otherwise live with every
-# later sibling masked from here on. Same C++20-boundary-literal
-# restriction as the ``#if`` form above.
-_PP_ELIF_CPLUSPLUS_LESS_THAN_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*elif[ \t]+" + _PP_CPLUSPLUS_LESS_THAN_CPP20_TAIL + rb"[ \t]*$"
-)
-# ``#if defined(__cplusplus)``/``#if defined __cplusplus`` (the
-# parenthesized *and* the equally-valid no-parens ``defined`` operator
-# spelling — Codex review, ninth round)/``#if __cplusplus`` (bare, no
-# comparison) are all semantically identical to ``#ifdef __cplusplus`` —
-# unconditionally true for every ``-std=`` this heuristic could pick — so
-# they must be excluded from the general guard pattern above the same way
-# ``#ifdef __cplusplus`` already is, not masked as if circular (Codex
-# review, eighth round). Anchored ``$`` so it only matches when this is
-# the *entire* condition: ``#if defined(__cplusplus) && __cplusplus >=
-# 202002L`` still has a genuine version comparison attached and stays
-# mask-worthy.
-_PP_DEFINED_CPLUSPLUS = (
-    rb"defined[ \t]*(?:\([ \t]*__cplusplus[ \t]*\)|[ \t]+__cplusplus)"
-)
-# ``#if defined(__cplusplus) && !defined(SWIG)`` (Codex review, twentieth
-# round) — a common real-world idiom wrapping a header's C++-only section
-# so a non-C++ consumer (SWIG, Doxygen, ...) doesn't see it, not a genuine
-# dialect/feature-test guard at all: the second conjunct doesn't reference
-# __cplusplus/__cpp_* so its truth value has nothing to do with the -std=
-# this heuristic picks, and this heuristic never *is* one of those non-C++
-# consumers, so the whole condition is unconditionally true here just like
-# the bare ``defined(__cplusplus)`` form above. The identifier is
-# excluded from matching __cplusplus/__cpp_* itself (a negative lookahead)
-# so a genuinely dialect-related second conjunct (``defined(__cplusplus)
-# && !defined(__cpp_concepts)``) is not swept in here and stays
-# mask-worthy, same as any other feature-test guard. Deliberately narrow
-# (single ``&&``, one negated macro) rather than a full expression parser,
-# matching this file's incremental-per-reported-case scope.
-_PP_NOT_DEFINED_UNRELATED_MACRO = (
-    rb"!defined[ \t]*(?:\([ \t]*(?!__cplusplus\b|__cpp_)\w+[ \t]*\)"
-    rb"|[ \t]+(?!__cplusplus\b|__cpp_)\w+)"
-)
-_PP_CPLUSPLUS_ALWAYS_TRUE_TAIL = (
-    rb"(?:"
-    + _PP_DEFINED_CPLUSPLUS
-    + rb"|__cplusplus"
-    + rb"|(?:"
-    + _PP_DEFINED_CPLUSPLUS
-    + rb")[ \t]*&&[ \t]*"
-    + _PP_NOT_DEFINED_UNRELATED_MACRO
-    + rb")"
-)
-_PP_IF_CPLUSPLUS_ALWAYS_TRUE_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*if[ \t]+" + _PP_CPLUSPLUS_ALWAYS_TRUE_TAIL + rb"[ \t]*$"
-)
-_PP_ELIF_CPLUSPLUS_ALWAYS_TRUE_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*elif[ \t]+" + _PP_CPLUSPLUS_ALWAYS_TRUE_TAIL + rb"[ \t]*$"
-)
-# The common shorthand spellings of a feature-test guard (Codex review,
-# seventh round): ``#ifdef __cpp_concepts`` is exactly as circular as
-# ``#if defined(__cpp_concepts)`` — masked for the same reason (the
-# feature-guarded arm is untrustworthy; its ``#else`` fallback is treated
-# as live, same as the general ``#if defined(__cpp_x)`` case).
-# ``#ifndef __cplusplus`` is different and still mask-worthy the same
-# way: castxml/clang always parse these headers in a C++-ish mode (see
-# ``_CPP_PATTERNS`` above), so that branch is never reachable regardless
-# of which ``-std=`` gets picked, exactly like ``#if 0``.
-# ``#ifdef __cplusplus`` is deliberately *not* included here — unlike a
-# version/feature-test comparison, it is unconditionally true for every
-# ``-std=`` this heuristic could pick (castxml always defines it), so its
-# content is a genuine, unconditional signal and must be scanned normally
-# rather than masked away.
-_PP_IFDEF_CPLUSPLUS_FEATURE_GUARD_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*ifdef[ \t]+__cpp_\w+\b"
-)
-_PP_IFNDEF_CPLUSPLUS_PATTERN = re.compile(rb"^[ \t]*#[ \t]*ifndef[ \t]+__cplusplus\b")
-# ``#ifdef __cplusplus`` itself (Codex review, twenty-first round): the
-# reasoning above ("unconditionally true for every -std= this heuristic
-# could pick") only holds once the language mode is *already* C++ --
-# ``mask_cplusplus_defined_guards`` (see the docstring below) is for the
-# one caller (deciding whether to treat an auto-detected header as C or
-# C++ at all) where that assumption doesn't hold: in C mode __cplusplus
-# is undefined, so this guard's content is exactly as unreachable there
-# as any other feature-test guard, and forcing C++ purely because C++20
-# syntax exists behind it would itself be the circular mistake this
-# masking exists to avoid.
-_PP_IFDEF_CPLUSPLUS_PATTERN = re.compile(rb"^[ \t]*#[ \t]*ifdef[ \t]+__cplusplus\b")
-# ``#ifndef __cplusplus``/``#if !defined(__cplusplus)`` (Codex review,
-# twenty-third round): the general mask-bucket's unconditional treatment
-# of these ("always parsed in a C++-ish mode, so never reachable") is the
-# same dialect-decision-only assumption as ``#ifdef __cplusplus`` above,
-# and just as wrong for ``mask_cplusplus_defined_guards`` -- in C mode
-# ``__cplusplus`` really is undefined, so this guarded arm (the C
-# fallback) is exactly the reachable one there, and the ``#else`` (C++-
-# only content) is the circular one instead. Opposite polarity from
-# ``#ifdef __cplusplus`` for the same reason ``#ifndef __cpp_x`` is
-# opposite from ``#ifdef __cpp_x``.
-_PP_NOT_DEFINED_CPLUSPLUS = rb"![ \t]*" + _PP_DEFINED_CPLUSPLUS
-_PP_IF_NOT_DEFINED_CPLUSPLUS_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*if[ \t]+" + _PP_NOT_DEFINED_CPLUSPLUS + rb"[ \t]*$"
-)
-_PP_ELIF_NOT_DEFINED_CPLUSPLUS_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*elif[ \t]+" + _PP_NOT_DEFINED_CPLUSPLUS + rb"[ \t]*$"
-)
-# ``#ifndef __cpp_x`` is the *negated* mirror of ``#ifdef __cpp_x`` —
-# logically ``#if defined(__cpp_x) ... [B] ... #else ... [A] ... #endif``
-# with the arms swapped, so it needs the opposite polarity, not the same
-# masking as ``#ifndef __cplusplus`` (Codex review, fourteenth round). The
-# ``#ifndef``-guarded arm here is the feature-*absent* content — the
-# portable/pre-C++20-safe fallback, exactly like the ``#else`` of the
-# positive form — and must be the one trusted as live; the ``#else`` here
-# is the feature-*present* content (typically the C++20-specific code,
-# circular for the same reason the positive form's guarded arm is) and
-# must be the one masked instead. Reusing the ``__cplusplus``-style
-# dead-first treatment for this pattern got the polarity backwards: it
-# masked the safe fallback and trusted the circular feature-present
-# ``#else`` arm, so a header whose *only* C++20 construct sat behind such
-# an ``#else`` forced ``-std=gnu++20`` purely because the guard's content
-# was there — which could then break an unrelated, genuinely pre-C++20
-# use of the same word as an ordinary identifier elsewhere in the header.
-_PP_IFNDEF_CPP_FEATURE_GUARD_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*ifndef[ \t]+__cpp_\w+\b"
-)
-# ``#if !defined(__cpp_x)``/``#if !defined __cpp_x`` (Codex review,
-# eighteenth round) is just the ``defined``-operator spelling of
-# ``#ifndef __cpp_x`` — same negated-mirror polarity, same reasoning as
-# the comment above. Without this, the general ``__cplusplus``/``__cpp_*``
-# guard pattern still matches it (it contains ``__cpp_x`` too) and applies
-# the *positive*-form polarity instead: masking the guarded arm (the
-# feature-absent, portable fallback) and trusting the ``#else`` (the
-# circular feature-present content) -- backwards, exactly the same bug the
-# ``#ifndef`` spelling above was fixed for. Anchored ``$`` so only a bare
-# negated feature-test counts as the whole condition, matching how the
-# other single-purpose guards here are deliberately narrow rather than a
-# full expression parser. Checked ahead of the general branch below, the
-# same way the ``#ifndef`` spelling doesn't need to be (it can never match
-# the general branch's ``#if``-only pattern), since this spelling starts
-# with ``#if`` and would otherwise be caught by it first.
-_PP_NOT_DEFINED_CPP_FEATURE_TAIL = (
-    rb"![ \t]*defined[ \t]*(?:\([ \t]*__cpp_\w+[ \t]*\)|[ \t]+__cpp_\w+)"
-)
-_PP_IF_NOT_DEFINED_CPP_FEATURE_GUARD_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*if[ \t]+" + _PP_NOT_DEFINED_CPP_FEATURE_TAIL + rb"[ \t]*$"
-)
-# Same ``#elif`` companion pattern this file already adds for every other
-# inverted-polarity ``#if`` spelling (the less-than-C++20 guard above),
-# for a chain that opens on something else and reaches the negated
-# feature-test guard as a later arm.
-_PP_ELIF_NOT_DEFINED_CPP_FEATURE_GUARD_PATTERN = re.compile(
-    rb"^[ \t]*#[ \t]*elif[ \t]+" + _PP_NOT_DEFINED_CPP_FEATURE_TAIL + rb"[ \t]*$"
-)
-
-
-def _strip_inactive_if_zero_blocks(
-    content: bytes,
-    *,
-    invert_dialect_fallback_guards: bool = True,
-    mask_cplusplus_defined_guards: bool = False,
-) -> bytes:
-    """Blank out unreachable arms of an ``#if 0``/``#if false`` chain.
-
-    A disabled compatibility stub (``#if 0\\nstruct consteval {};\\n#endif``)
-    must not shadow a genuine ``consteval``/``constinit``/``concept`` keyword
-    used as an identifier elsewhere in the header (Codex review) — the
-    ``*_type_shadowed`` scan otherwise sees the inactive type definition and
-    wrongly treats a real C++20 declaration as ambiguous. Line count is
-    preserved so this can be composed with the rest of the shadow scan.
-
-    Only *unreachable* arms are masked. ``#if 0``/``#elif 0``/``#elif
-    false`` are permanently unreachable, exactly like their guard says
-    (Codex review, second round). An ``#else``/``#elif`` whose own
-    condition isn't a recognizable ``0``/``1``/``false``/``true`` literal
-    has a condition this heuristic can't evaluate, so it's treated as
-    possibly-reachable (unmasked) — the conservative direction, since it
-    can only request C++20 mode when not strictly needed, never the
-    reverse. But once an ``#elif 1``/``#elif true`` arm is seen, it is
-    *unconditionally* reachable in every build configuration, which makes
-    every later sibling arm in that same chain unconditionally
-    *unreachable* — masking must resume for them too (Codex review, third
-    round), not stay lifted for the rest of the chain the way it correctly
-    does after a merely-unevaluated condition. The same applies when the
-    chain *opens* on an unconditionally-true literal (``#if 1``/``#if
-    true``, Codex review, tenth round): that arm is scanned normally like
-    any other unmasked arm, but the chain is entered and immediately
-    settled so a dead ``#else``/``#elif`` sibling doesn't fall through to
-    plain pass-through scanning — a real C++20 construct sitting only in
-    that dead sibling would otherwise force ``-std=gnu++20`` onto a header
-    whose live code never needed it, breaking any active use of a
-    soon-to-be-reserved word as an ordinary identifier.
-
-    An ``#if``/``#elif`` guarded on ``__cplusplus`` or a standard
-    feature-test macro (``__cpp_concepts``, ``__cpp_consteval``, ...) is
-    masked the same way as ``#if 0`` (Codex review, sixth round) — not
-    because it's provably false, but because it's circular: whichever
-    ``-std=`` this heuristic ends up choosing is exactly what decides
-    whether the guard is reachable, so its content must never be allowed
-    to drive that same choice. Forcing C++20 mode purely because a
-    ``#if __cplusplus >= 202002L`` block contains C++20 syntax doesn't
-    just needlessly activate that block — passing ``-std=gnu++20`` turns
-    every *unguarded* pre-C++20 use of ``consteval``/``constinit``/
-    ``concept``/``requires`` as an ordinary identifier elsewhere in the
-    same header into a reserved-word parse error too.
-
-    A nested ``#if`` inside a *reachable* arm gets the exact same
-    treatment, recursively, via an explicit per-depth stack rather than a
-    flat "opaque nested" counter (Codex review, eleventh round). A flat
-    counter that just tracks depth to find the matching ``#endif`` cannot
-    tell a nested ``#if 0`` compatibility stub from nested live code, so a
-    dead nested construct inside an otherwise-live arm previously leaked
-    into the shadow/requirements scan (or, symmetrically, a live nested
-    construct inside a dead-sibling ancestor was wrongly treated as
-    reachable). Each stack frame independently recognizes zero/true/
-    feature-guard conditions the same way the top level does; a frame
-    whose own condition isn't recognizable stays fully unmasked (pass
-    through, the same conservative default the top level already applies
-    to an unrecognized ``#if``), but any frame with an unreachable
-    *ancestor* is unconditionally unreachable regardless of its own
-    condition.
-
-    ``invert_dialect_fallback_guards`` (default ``True``) controls the
-    ``__cplusplus < N``/``#ifndef __cpp_x``/``#if !defined(__cpp_x)``
-    inverted-polarity treatment described above. It must be **disabled**
-    (``False``) for masking content that feeds the *shadow*-name scan
-    (Codex review, nineteenth round): that scan asks "does 'concept' name
-    an ordinary pre-C++20 compatibility type *in code that would still be
-    reachable if C++20 were chosen*" — the inverted polarity is exactly
-    backwards for that question, since it treats the guarded arm of e.g.
-    ``#if __cplusplus < 202002L`` as live specifically *because* it is a
-    pre-C++20 fallback shim, i.e. content that goes away once C++20 is
-    chosen. Left enabled there, a ``struct concept {};`` compatibility
-    shim confined entirely to such a guard would wrongly shadow a genuine
-    ``concept`` declaration in another header of the same aggregate,
-    permanently keeping the whole translation unit off ``-std=gnu++20``
-    even though the shim itself would never coexist with that declaration
-    under any dialect choice. The *requirements* scan (deciding whether
-    unconditionally-reachable code needs C++20) is unaffected and keeps
-    the default ``True``.
-
-    ``mask_cplusplus_defined_guards`` (default ``False``) additionally
-    masks ``#if __cplusplus``/``#ifdef __cplusplus``/``#if
-    defined(__cplusplus)`` (bare or with the SWIG-idiom ``&&
-    !defined(X)`` extension) instead of treating them as an
-    unconditionally-true, unmasked signal. Those forms are exempted by
-    default because *given the header is already being parsed as C++*,
-    ``__cplusplus`` really is always defined — correct for deciding the
-    C++ **dialect** (which ``-std=gnu++NN``). It must be **enabled**
-    (``True``) for the one caller that instead uses this scan's result to
-    decide whether to treat an auto-detected header as C or C++ **at
-    all** (Codex review, twenty-first round): in C mode ``__cplusplus``
-    is undefined, so a ``consteval``/``concept``/... construct confined
-    to such a guard is not actually reachable there, and letting it force
-    C++ mode is the exact same circularity this function exists to
-    prevent elsewhere — worse, it then turns an *active*, unguarded use
-    of the same word as an ordinary C identifier elsewhere in the header
-    into a reserved-word parse error once C++20 mode is wrongly forced.
-    """
-    lines = content.split(b"\n")
-    out: list[bytes] = []
-    # One entry per currently-open #if, innermost last. Each entry is
-    # [recognized, masking, settled]:
-    #   recognized: this level's own condition matched one of the
-    #     zero/true/feature-guard patterns, so its later #elif/#else arms
-    #     are interpreted the same way the top level interprets them.
-    #   masking: the arm *currently open at this level* should be
-    #     blanked, judged purely by this level's own condition (ignoring
-    #     ancestors).
-    #   settled: an unconditionally-true arm already fired at this level,
-    #     so every later sibling arm here is unconditionally unreachable.
-    stack: list[list[bool]] = []
-
-    def unreachable() -> bool:
-        # Unreachable if this level or any ancestor is masking — a dead
-        # ancestor arm makes everything nested inside it dead too,
-        # regardless of the nested content's own condition.
-        return any(frame[1] for frame in stack)
-
-    for raw_line in lines:
-        # A CRLF source (or a CRLF-normalizing text-mode write, e.g. the
-        # test suite's Path.write_text() on Windows) leaves a trailing "\r"
-        # on every line after splitting on "\n" alone — which the trailing
-        # "$" in these patterns then fails to match, since "\r" isn't in
-        # their "[ \t]*" tail (Windows CI regression). Strip it before
-        # matching; harmless on genuine LF input.
-        line = raw_line[:-1] if raw_line.endswith(b"\r") else raw_line
-
-        if _PP_IF_OPEN_PATTERN.match(line):
-            if unreachable():
-                # Already inside dead code — this nested chain's own
-                # condition is irrelevant either way; track it only to
-                # find its matching #endif.
-                stack.append([False, True, True])
-                out.append(b"")
-                continue
-            if (
-                invert_dialect_fallback_guards
-                and _PP_IF_CPLUSPLUS_LESS_THAN_PATTERN.match(line)
-            ):
-                # Opposite polarity from the general __cplusplus-guard
-                # branch below: a less-than comparison's guarded arm is
-                # the pre-C++20-safe fallback, so it's scanned live and
-                # immediately settled so any #else (feature present,
-                # circular) stays masked -- checked ahead of the general
-                # branch since it would otherwise also match. Skipped
-                # entirely when invert_dialect_fallback_guards is False
-                # (shadow-scan callers), falling through to the general
-                # branch below instead -- see the docstring.
-                stack.append([True, False, True])
-                out.append(b"")
-                continue
-            if (
-                invert_dialect_fallback_guards
-                and _PP_IF_NOT_DEFINED_CPP_FEATURE_GUARD_PATTERN.match(line)
-            ):
-                # Same inverted polarity as #ifndef __cpp_x below --
-                # checked ahead of the general branch since this spelling
-                # (unlike #ifndef) starts with "#if" and would otherwise
-                # be caught there first. Same shadow-scan exception as
-                # above.
-                stack.append([True, False, True])
-                out.append(b"")
-                continue
-            if mask_cplusplus_defined_guards and (
-                _PP_IFNDEF_CPLUSPLUS_PATTERN.match(line)
-                or _PP_IF_NOT_DEFINED_CPLUSPLUS_PATTERN.match(line)
-            ):
-                # Opposite polarity from the general branch below's
-                # unconditional #ifndef __cplusplus/!defined(__cplusplus)
-                # treatment -- checked ahead of it since it would
-                # otherwise also match. See the docstring above
-                # _PP_NOT_DEFINED_CPLUSPLUS.
-                stack.append([True, False, True])
-                out.append(b"")
-                continue
-            if (
-                _PP_IF_ZERO_PATTERN.match(line)
-                or _PP_IFDEF_CPLUSPLUS_FEATURE_GUARD_PATTERN.match(line)
-                or _PP_IFNDEF_CPLUSPLUS_PATTERN.match(line)
-                or (
-                    not invert_dialect_fallback_guards
-                    and _PP_IFNDEF_CPP_FEATURE_GUARD_PATTERN.match(line)
-                )
-                or (
-                    mask_cplusplus_defined_guards
-                    and _PP_IFDEF_CPLUSPLUS_PATTERN.match(line)
-                )
-                or (
-                    _PP_IF_CPLUSPLUS_GUARD_PATTERN.match(line)
-                    and (
-                        mask_cplusplus_defined_guards
-                        or not _PP_IF_CPLUSPLUS_ALWAYS_TRUE_PATTERN.match(line)
-                    )
-                )
-            ):
-                stack.append([True, True, False])
-                out.append(b"")
-                continue
-            if _PP_IF_TRUE_PATTERN.match(line) or (
-                not mask_cplusplus_defined_guards
-                and (
-                    _PP_IFDEF_CPLUSPLUS_PATTERN.match(line)
-                    or _PP_IF_CPLUSPLUS_ALWAYS_TRUE_PATTERN.match(line)
-                )
-            ):
-                # ``#if defined(__cplusplus)``/``#ifdef __cplusplus``/bare
-                # ``#if __cplusplus`` (Codex review, twenty-second round):
-                # tracked as *settled*-true here, the same way ``#if 1``/
-                # ``#if true`` already is, not left as an unrecognized
-                # chain (whose #else/#elif arms then get scanned right
-                # along with this one, unmasked). Without this, a dual
-                # C/C++ header's ``#else`` (C-only) fallback -- e.g. a
-                # ``struct concept {};`` compatibility shim -- stayed
-                # visible to the shadow-name scan alongside a genuine
-                # C++20 declaration in the *active* ``#if`` arm, wrongly
-                # shadowing it. Elif-level already settles the equivalent
-                # ``#elif`` spelling this same way; this closes the gap
-                # for it as the chain's *opening* condition. Skipped when
-                # mask_cplusplus_defined_guards is True (the caller
-                # already masked this guard's own content in the general
-                # branch above instead, and never reaches this line).
-                stack.append([True, False, True])
-                out.append(b"")
-                continue
-            if (
-                invert_dialect_fallback_guards
-                and _PP_IFNDEF_CPP_FEATURE_GUARD_PATTERN.match(line)
-            ):
-                # Opposite polarity from the branch above: the
-                # #ifndef-guarded arm (feature absent) is the trustworthy
-                # one here, so it's scanned live and immediately settled
-                # so any #else (feature present, circular) stays masked.
-                # Skipped for a shadow-scan caller -- already folded into
-                # the general mask branch above instead.
-                stack.append([True, False, True])
-                out.append(b"")
-                continue
-            # Unrecognized condition — same conservative pass-through the
-            # top level always applied to an #if it can't evaluate.
-            stack.append([False, False, False])
-            out.append(line)
-            continue
-
-        if stack and _PP_ENDIF_PATTERN.match(line):
-            stack.pop()
-            out.append(b"" if unreachable() else line)
-            continue
-
-        if stack:
-            frame = stack[-1]
-            ancestors_unreachable = any(f[1] for f in stack[:-1])
-            if ancestors_unreachable:
-                out.append(b"")
-                continue
-            if not frame[0]:
-                # This level's own #if condition was unrecognized — its
-                # #elif/#else arms are never specially interpreted,
-                # exactly like the top level's unrecognized-chain
-                # pass-through.
-                out.append(line)
-                continue
-            if (
-                invert_dialect_fallback_guards
-                and _PP_ELIF_CPLUSPLUS_LESS_THAN_PATTERN.match(line)
-            ):
-                # Same inverted polarity as the #if-opening case, checked
-                # ahead of the general branch below since it would
-                # otherwise also match. Same shadow-scan exception as the
-                # #if-opening case: falls through to the general branch
-                # (mask) when invert_dialect_fallback_guards is False.
-                frame[1] = frame[2]
-                frame[2] = True
-                out.append(b"" if frame[1] else line)
-                continue
-            if (
-                invert_dialect_fallback_guards
-                and _PP_ELIF_NOT_DEFINED_CPP_FEATURE_GUARD_PATTERN.match(line)
-            ):
-                # Same inverted polarity as #if !defined(__cpp_x) above,
-                # checked ahead of the general branch since it would
-                # otherwise also match. Same shadow-scan exception.
-                frame[1] = frame[2]
-                frame[2] = True
-                out.append(b"" if frame[1] else line)
-                continue
-            if (
-                mask_cplusplus_defined_guards
-                and _PP_ELIF_NOT_DEFINED_CPLUSPLUS_PATTERN.match(line)
-            ):
-                # Opposite polarity from the general branch below's
-                # unconditional !defined(__cplusplus) treatment, checked
-                # ahead of it since it would otherwise also match. Same
-                # reasoning as the #if-opening case above.
-                frame[1] = frame[2]
-                frame[2] = True
-                out.append(b"" if frame[1] else line)
-                continue
-            if _PP_ELIF_ZERO_PATTERN.match(line) or (
-                _PP_ELIF_CPLUSPLUS_GUARD_PATTERN.match(line)
-                and (
-                    mask_cplusplus_defined_guards
-                    or not _PP_ELIF_CPLUSPLUS_ALWAYS_TRUE_PATTERN.match(line)
-                )
-            ):
-                frame[1] = True
-                out.append(b"")
-                continue
-            if _PP_ELIF_TRUE_PATTERN.match(line) or (
-                not mask_cplusplus_defined_guards
-                and _PP_ELIF_CPLUSPLUS_ALWAYS_TRUE_PATTERN.match(line)
-            ):
-                frame[1] = frame[2]
-                frame[2] = True
-                out.append(b"" if frame[1] else line)
-                continue
-            if _PP_ELIF_PATTERN.match(line) or _PP_ELSE_PATTERN.match(line):
-                frame[1] = frame[2]
-                out.append(b"" if frame[1] else line)
-                continue
-            out.append(b"" if frame[1] else line)
-            continue
-
-        out.append(line)
-    return b"\n".join(out)
-
-
-@dataclass(frozen=True)
-class Cpp20Requirement:
-    """A single structural C++20 construct found while scanning headers."""
-
-    reason: Literal[
-        "concept-declaration",
-        "requires-expression",
-        "requires-clause",
-        "constrained-template-parameter",
-        "custom-constrained-auto-parameter",
-        "abbreviated-function-template-parameter",
-        "consteval-declaration",
-        "constinit-declaration",
-    ]
-    path: str
-    line: int
-
-
-def _find_cpp20_requirements(
-    header_paths: list[Path], *, for_language_mode_decision: bool = False
-) -> list[Cpp20Requirement]:
-    """Scan *header_paths* for structural C++20 syntax, with reasons/locations.
-
-    Conservative and directive/literal/comment-aware: only definition-site
-    syntax in actual code counts, never the same keywords appearing inside a
-    preprocessor diagnostic message, a string/char literal, or a comment.
-
-    Two passes over *header_paths*, not one: *header_paths* is the whole
-    aggregate header set castxml/clang parses together as a single
-    translation unit, so a pre-C++20 compatibility type shadowing
-    "concept"/"requires"/"consteval"/"constinit" (Codex review) can live in
-    one file (a shared ``compat.hpp``) while the ambiguous bare use sits in
-    another (``api.hpp``) — a per-file shadow check would see only
-    ``api.hpp``'s own content and miss it, wrongly forcing C++20 mode on
-    the whole aggregate (Codex review, sixth round). The first pass reads
-    and preprocesses every file and ORs each shadow flag across all of
-    them; the second reuses that same preprocessed content to run the
-    per-line scan against the now aggregate-wide flags.
-
-    ``for_language_mode_decision`` (default ``False``) must be set by the
-    one caller that uses this scan's result to decide whether an
-    auto-detected header is C or C++ **at all**, not just which C++
-    dialect (Codex review, twenty-first round): it masks
-    ``#if __cplusplus``/``#ifdef __cplusplus``/``defined(__cplusplus)``
-    -guarded content (see ``_strip_inactive_if_zero_blocks``'s
-    ``mask_cplusplus_defined_guards``) instead of treating it as an
-    always-true, unmasked signal — that assumption only holds once C++
-    mode is already chosen, and is exactly backwards for the decision of
-    whether to choose it in the first place.
-    """
-    per_file: list[tuple[Path, bytes]] = []
-    concept_type_shadowed = False
-    requires_type_shadowed = False
-    consteval_type_shadowed = False
-    constinit_type_shadowed = False
-    for p in header_paths:
-        try:
-            content = p.read_bytes()
-        except OSError:
-            continue
-        # Blank raw string literals first — their body can contain arbitrary
-        # quotes/backslashes that would otherwise confuse the ordinary
-        # string-literal stripper below.
-        content = _strip_raw_strings(content)
-        # Blank string/char literals first so a literal containing comment-like
-        # text ("/* not a comment */") is never mistaken for a real comment.
-        # Backslash-newline-continuation-tolerant (Codex review): a literal
-        # split across a continuation must not leave its trapped text — e.g.
-        # a fake "struct concept {};" inside an error message — unblanked
-        # and visible to the shadow-name scan below.
-        content = _strip_literals_crossing_continuations(content)
-        # Strip real block comments, but preserve the embedded newline count so
-        # later-reported line numbers stay accurate for code following a
-        # multi-line comment (CodeRabbit review).
-        content = re.sub(
-            rb"/\*.*?\*/",
-            lambda m: b"\n" * m.group(0).count(b"\n"),
-            content,
-            flags=re.DOTALL,
-        )
-        # Whether this header defines "concept" as an ordinary type name
-        # anywhere (Codex review, fourth round) — a pre-C++20 header can
-        # declare a type literally named "concept" and initialize a
-        # variable template of that type via *any* expression convertible
-        # to it (not just a brace-init-list, which only covered the
-        # aggregate-init case), so no per-initializer-shape check can be
-        # complete. Once "concept" is confirmed to name a real type
-        # anywhere in the aggregate, every bare "concept NAME = ..." match
-        # in it is ambiguous and treated as non-genuine — see
-        # ``_looks_like_genuine_concept``. Checked against a *separate*,
-        # additionally "//"-line-comment-stripped copy of the content
-        # (raw strings/literals/block comments are already blanked in
-        # ``content`` itself at this point, but not "//" comments, which
-        # are only stripped per-logical-line further below) — a
-        # "// struct concept {};" comment must never make a *real*
-        # concept declaration elsewhere in the header look ambiguous
-        # (Codex review, fifth round).
-        # Also strip out #if 0/#if false regions before any of the three
-        # shadow scans below — a disabled compatibility stub must not shadow
-        # a genuine keyword used elsewhere in the header (Codex review).
-        content_no_double_slash_comments = re.sub(rb"//[^\n]*", b"", content)
-        content_no_line_comments = _strip_inactive_if_zero_blocks(
-            content_no_double_slash_comments,
-            mask_cplusplus_defined_guards=for_language_mode_decision,
-        )
-        per_file.append((p, content_no_line_comments))
-        # A *separate* masking pass, not the one above, feeds the shadow
-        # scans below: invert_dialect_fallback_guards=False (Codex review,
-        # nineteenth round) — the shadow scan asks "does 'concept' name an
-        # ordinary type in code still reachable if C++20 were chosen", so a
-        # ``struct concept {};`` compatibility shim confined to e.g.
-        # ``#if __cplusplus < 202002L`` (content that specifically goes
-        # away once C++20 is chosen) must not count, unlike for the
-        # requirements scan above where that same guarded arm is the
-        # unconditionally-relevant one. See the docstring of
-        # _strip_inactive_if_zero_blocks.
-        content_for_shadow_scan = _strip_inactive_if_zero_blocks(
-            content_no_double_slash_comments, invert_dialect_fallback_guards=False
-        )
-        concept_type_shadowed = concept_type_shadowed or bool(
-            _CONCEPT_AS_TYPE_NAME_PATTERN.search(content_for_shadow_scan)
-        )
-        # Same reasoning as concept_type_shadowed above, for "requires" used
-        # as an ordinary pre-C++20 type name (Codex review).
-        requires_type_shadowed = requires_type_shadowed or bool(
-            _REQUIRES_AS_TYPE_NAME_PATTERN.search(content_for_shadow_scan)
-        )
-        # Same reasoning and "//"-comment-stripped-copy caveat as
-        # concept_type_shadowed above, for "consteval"/"constinit" used as
-        # an ordinary pre-C++20 type name (Codex review, third round).
-        consteval_type_shadowed = consteval_type_shadowed or bool(
-            _CONSTEVAL_AS_TYPE_NAME_PATTERN.search(content_for_shadow_scan)
-        )
-        constinit_type_shadowed = constinit_type_shadowed or bool(
-            _CONSTINIT_AS_TYPE_NAME_PATTERN.search(content_for_shadow_scan)
-        )
-
-    found: list[Cpp20Requirement] = []
-    for p, _content_no_line_comments in per_file:
-        # Scan the same #if-0-stripped content the shadow checks above use —
-        # a genuine consteval/constinit/concept/requires construct written
-        # only inside a disabled #if 0 block must not itself mark the header
-        # as needing C++20 (Codex review): it's never actually compiled.
-        logical_lines = _iter_logical_lines(_content_no_line_comments)
-        n = len(logical_lines)
-        # Last non-blank line's own (un-extended) code, tracked across
-        # iterations — lets a concept-declaration candidate look backward
-        # for its template<...> header when "concept" itself starts a line
-        # (see _looks_like_genuine_concept).
-        prev_nonblank_code = b""
-        for i, (start_no, logical) in enumerate(logical_lines):
-            if _is_preprocessor_directive(logical):
-                continue
-            code = _strip_literals_joined(logical)
-            code = code.split(b"//")[0]
-            # A bare "requires"/"concept"/"consteval"/"constinit" trailing at
-            # the end of a line (no parameter list/brace/name yet, or for
-            # consteval/constinit no declarator following at all) means the
-            # construct's continuation landed on a following physical line
-            # with no backslash join in between — the per-line scan
-            # otherwise never sees the two halves together (Codex review;
-            # the same gap applies symmetrically to "concept" split before
-            # its name, not just "requires" split before its "("/"{"/
-            # constraint, and equally to consteval/constinit split before
-            # their own declarator, e.g. ``consteval\nint f();`` — Codex
-            # review, second round). Pull in subsequent non-directive lines
-            # (bounded, so a stray trailing keyword in unrelated code can't
-            # scan unboundedly) until the bare-trailing condition no longer
-            # holds.
-            lookahead = code
-            j = i
-            lookahead_budget = 5
-            while (
-                lookahead_budget > 0
-                and re.search(
-                    rb"\b(?:requires|concept|consteval|constinit)\s*$",
-                    lookahead.rstrip(),
-                )
-                and j + 1 < n
-                and not _is_preprocessor_directive(logical_lines[j + 1][1])
-            ):
-                j += 1
-                nxt = _strip_literals_joined(logical_lines[j][1]).split(b"//")[0]
-                lookahead += b"\n" + nxt
-                lookahead_budget -= 1
-            concept_match = _CPP20_CONCEPT_PATTERN.search(lookahead)
-            requires_expr_match = _CPP20_REQUIRES_EXPR_PATTERN.search(lookahead)
-            if concept_match and _looks_like_genuine_concept(
-                lookahead, concept_match, prev_nonblank_code, concept_type_shadowed
-            ):
-                found.append(Cpp20Requirement("concept-declaration", str(p), start_no))
-            elif requires_expr_match and not _looks_like_requires_declarator(
-                lookahead, requires_expr_match, prev_nonblank_code
-            ):
-                found.append(Cpp20Requirement("requires-expression", str(p), start_no))
-            elif (
-                clause_match := _CPP20_REQUIRES_CLAUSE_PATTERN.search(lookahead)
-            ) and _looks_like_genuine_requires_clause(
-                lookahead,
-                clause_match.start(),
-                prev_nonblank_code,
-                requires_type_shadowed,
-            ):
-                found.append(Cpp20Requirement("requires-clause", str(p), start_no))
-            elif _has_constrained_param_syntax(lookahead):
-                found.append(
-                    Cpp20Requirement("constrained-template-parameter", str(p), start_no)
-                )
-            elif _has_custom_constrained_auto_param(lookahead, prev_nonblank_code):
-                found.append(
-                    Cpp20Requirement(
-                        "custom-constrained-auto-parameter", str(p), start_no
-                    )
-                )
-            elif _has_abbreviated_unconstrained_auto_param(lookahead):
-                found.append(
-                    Cpp20Requirement(
-                        "abbreviated-function-template-parameter", str(p), start_no
-                    )
-                )
-            elif not consteval_type_shadowed and _CPP20_CONSTEVAL_PATTERN.search(
-                lookahead
-            ):
-                found.append(
-                    Cpp20Requirement("consteval-declaration", str(p), start_no)
-                )
-            elif not constinit_type_shadowed and _CPP20_CONSTINIT_PATTERN.search(
-                lookahead
-            ):
-                found.append(
-                    Cpp20Requirement("constinit-declaration", str(p), start_no)
-                )
-            if code.strip():
-                prev_nonblank_code = code
-    return found
-
-
-def _detect_cpp20_headers(
-    header_paths: list[Path], *, for_language_mode_decision: bool = False
-) -> bool:
-    """Return True if any header contains structural C++20-only syntax
-    (concept/requires, a constrained or abbreviated function template
-    parameter, ``consteval``, or ``constinit``).
-
-    Used to decide whether to pass ``-std=gnu++20`` to castxml. castxml's
-    default standard is whatever the underlying compiler defaults to
-    (usually C++17 on modern gcc), which does not accept ``concept``
-    declarations. This detection is conservative: only definition-site
-    syntax counts, not the keyword in arbitrary text — see
-    ``_find_cpp20_requirements`` for the directive/literal/comment-aware scan.
-
-    ``for_language_mode_decision`` — see ``_find_cpp20_requirements``, to
-    which it's passed through unchanged. Must be set by the one caller
-    that uses this result to decide C vs. C++ auto-detection itself, not
-    just the C++ dialect.
-    """
-    return bool(
-        _find_cpp20_requirements(
-            header_paths, for_language_mode_decision=for_language_mode_decision
-        )
-    )
 
 
 def _detect_cpp_headers(
